Apolar-Molecules/pyproject.toml
[tool.poetry]
name = "Apolar-Molecules"
version = "0.1.0"
description = ""
authors = ["Miguel Senra"]

[tool.poetry.dependencies]
python = "^3.8"
dvc = "^2.10.0"
hydra-core = "^1.1.1"
pdoc3 = "^0.10.0"
notebook = "^6.4.10"
pre-commit = "^2.19.0"
numpy = "^1.24.3"
jupytext = "^1.14.5"
datasist = "^1.5.3"
scikit-learn = "^1.2.2"
xgboost = "^1.7.5"
scikit-optimize = "^0.9.0"
jupyter-contrib-nbextensions = "^0.7.0"
pandas = "^2.0.1"
ipykernel = "^6.22.0"
plotly-express = "0.4.0"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build   
)/ 
'''
